城市管理界面更新
逻辑层接口增加了一个

+++ 3_Coding/TransitSys/client/src/main/java/edu/nju/umr/vo/CitiesVO.java
@@ -3,9 +3,9 @@ package edu.nju.umr.vo;
 public class CitiesVO {
 	private String city1;
 	private String city2;
-	private int distance;
+	private double distance;
 	private double price;
-	public CitiesVO(String city1,String city2,int distance,double price) {
+	public CitiesVO(String city1,String city2,double distance,double price) {
 		// TODO 自动生成的构造函数存根
 		this.city1 = city1;
 		this.city2 = city2;
@@ -18,7 +18,7 @@ public class CitiesVO {
 	public String getCity2() {
 		return city2;
 	}
-	public int getDistance() {
+	public double getDistance() {
 		return distance;
 	}
 	public double getPrice() {
